messages-service/requirements.txt
flask==2.2.3
requests==2.28.1
kafka-python==2.0.2
Werkzeug==2.2.2
